+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 2.3.0
 * Overwrite existing tables in the metastore
-* Add functions module and provide switch_case column generation
+* Add functions module and provide switch_case column generation and multijoin
 
 # 2.2.1
 * `spark.sql.shuffle.partitions` in `SparklyTest` should be set to string,